A tiny spelling correct. Correct spelling is a lossed art.
